config.py
# config.py
import os
import re
import yaml


def _resolve_env_vars(value):
    """Replace ${VAR_NAME} patterns with environment variable values."""
    if isinstance(value, str):
        return re.sub(
            r"\$\{(\w+)\}",
            lambda m: os.environ.get(m.group(1), m.group(0)),
            value,
        )
    if isinstance(value, dict):
        return {k: _resolve_env_vars(v) for k, v in value.items()}
    if isinstance(value, list):
        return [_resolve_env_vars(item) for item in value]
    return value


def load_config(path: str) -> dict:
    """Load config.yaml, resolve env vars, return dict."""
    with open(path) as f:
        raw = yaml.safe_load(f)
    return _resolve_env_vars(raw)
